COMMON: Add support for arrays & limited custom deleters to smart pointers
This enables ScopedPtr to be used for all cases where raw pointers
are currently used to manage owned memory, and brings ScopedPtr
closer to std::unique_ptr in terms of functionality. DisposablePtr
also comes along for the ride.

Custom deleters of ScopedPtr are not currently fully conforming to
std::unique_ptr for the sake of simplicity of implementation.
Unlike in the standard library, plain functions and lvalue
references are not supported, nor may custom deleters be passed
to the constructor at runtime. This can be improved in the future,
if necessary, with zero overhead, by creating a Pair class that
uses the Empty Base Optimization idiom to avoid extra storage
overhead of the deleter instance when it is not needed, as in
typical standard library implementations, plus some additional
type traits to support the necessary metaprogramming for the
different type overloads.

Array types in SharedPtr also do not conform to C++17 which checks
for convertability between the argument and an array type, but
this should be an OK compromise for now.