SCI: Fix incorrect use of Sci1Sound::id
In SSCI this is used only for stopping sounds and save/load
operations. The rest of the time, the identity of the sound is
taken from the sound object when it is needed.

Fixes crashes when clicking a button on the Shivers main menu,
among probably a lot of other things.