SCI: Ignore invalid sound objects in save games
The pq4-cd.001 save on Trac#10392 has a saved sound object which
points to a VM object which does not exist and it has no resource
number or any state information. Trying to restore this object
causes a write error in the VM because the associated nodePtr
object does not exist.