SCI: Unconfuse different resource source types
There are two kinds of resource sources, index sources and data
sources. The root ResourceSource was acting like a null-indexer and
a VolumeResourceSource loader, which was wrong. Now, the resource
system still tries to scan from data-only sources, but at least it
it is not confusing any more about which are data sources and which
are index sources.