SCI32: Remove busted GfxCache code from SCI32 & fix GfxFont
GfxCache is really broken. Once the number of cached fonts goes
over the cache limit it just deletes all of them, so you can't
retain any references to anything in it since it could free the
memory at any time, and it periodically dumps the entire cache for
no reason. Furthermore, the parts of a font which are in any way
heavyweight is the resource data, and that is properly cached
already by ResourceManager.

This also fixes a memory leak in GfxFontSjis, fixes GfxFont so it
does not hold a reference to screen unnecessarily (which is only
ever used when it is actually time to draw to the screen and
the caller always has the screen reference already), fixes GfxFont
so it does not double-release a held Resource if it is copy-
constructed, and deduplicates rendering code. It's not perfect,
but it's better than it used to be.