SCI32S2: Rejigger Shivers 2 initialisation
This changes where two-phase initialisation of objects is needed.
Since the resource manager needs to initialise before almost
anything else can initialise because it sets the SCI version,
and the resource manager cannot do that until after
`Engine::initializePath` is called, nothing can really be a member
of the Engine subclass without requiring two-phase initialisation
for everything, which is undesirable. There is otherwise no real
reason why things cannot just be members of a single class, so
that is mostly what is being done now, with the exception that the
kernel and game are made separate to match more closely the
architecture of SSCI.

I could decide this is all a bad idea again later, we’ll see.